requirements.txt
apache-airflow==2.10.5
box==0.1.5
duckdb==1.1.3
Fabric==3.2.2
FLAML==2.1.1
invoke==2.2.0
joblib==1.3.2
lightgbm==4.6.0
lightning==2.4.0
loguru==0.7.2
matplotlib==3.10.1
miceforest==6.0.3
numba==0.59.1
numpy==2.2.3
pacmap==0.8.0
pandas==2.2.3
peft==0.13.2
polars==1.24.0
pytest==8.3.2
python_box==7.3.2
PyYAML==6.0.1
PyYAML==6.0.2
rich==13.9.4
schedulefree==1.4
scikit_learn==1.3.2
scikit_learn_intelex==2024.6.0
scipy==1.15.2
seaborn==0.13.2
sentence_transformers==3.2.1
tableone==0.9.1
torch==2.6.0
tqdm==4.66.4
transformers==4.49.0
